make specific js files for test shared functions
